Create Phidias.py
This is my first initialized version the FizzBuzz challenge. First, we create the for loop to use numbers from 1 to 50. Every time the for loop iterates, we start with a blank output string. Then we test whether index, the iterated number, is divisible by 3 or divisible by 5. If our index is divisible by 3, it stores "RAM" to the output variable. If it's divisible by 5, it stores "Hacks" to the output variable. If it is not divisible by both, it will store the index to the output variable. At the end of that for loop, it will then print out the output variable.